+++ packages/components/src/Form/Inputs/InputProps.ts
@@ -24,11 +24,14 @@
 
  */
 
+import pick from 'lodash/pick'
 import { CompatibleHTMLProps } from '@looker/design-tokens'
 import { ValidationType } from '../ValidationMessage'
 
 export interface InputProps extends CompatibleHTMLProps<HTMLInputElement> {
   validationType?: ValidationType
+  'data-autofocus'?: string
+  'data-testid'?: string
 }
 
 export interface InputTextTypeProps {
@@ -92,3 +95,19 @@ export const inputPropKeys = [
   'aria-describedby',
   'aria-labelledby',
 ]
+/**
+ * Adds data-autofocus attribute for use with focus-trap (Popover & Dialog)
+ * Can't use the autofocus dom attribute b/c React does not pass it down
+ * https://github.com/facebook/react/issues/11851
+ */
+export const getAutoFocusProps = (autoFocus?: boolean) => {
+  return autoFocus ? { autoFocus, 'data-autofocus': 'true' } : {}
+}
+
+export const pickInputProps = <T extends { autoFocus?: boolean }>({
+  autoFocus,
+  ...props
+}: T) => {
+  const inputProps = pick(props, inputPropKeys)
+  return { ...getAutoFocusProps(autoFocus), ...inputProps }
+}