+++ packages/components/src/Form/Inputs/Select/SelectMulti.tsx
@@ -74,7 +74,6 @@ const SelectMultiComponent = forwardRef(
       options,
       disabled,
       autoFocus,
-      'data-autofocus': dataAutofocus,
       isFilterable = false,
       placeholder,
       onFilter,
@@ -127,7 +126,6 @@ const SelectMultiComponent = forwardRef(
           {...ariaProps}
           disabled={disabled}
           autoFocus={autoFocus}
-          data-autofocus={dataAutofocus}
           placeholder={placeholder}
           removeOnBackspace={removeOnBackspace}
           validationType={props.validationType}